Set up Return to Computer
Added in Behaviors for Sonic Mail and Scenario2To1 Transitioner to bring the Player back to Scenario 1 once all 4 Instruments are grown. A "Victory Window" now appears that allows Players to hear the victory song and send the Mail off. Also fixed a bug that was allowing duplicate Plants to count toward finishing the Mail.